requirements.txt
pyautogui>=0.9.54
pygetwindow>=0.0.9
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
